Fix what has been broken by the merge

+++ backend/app/models/out_of_stock_item.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, Enum, Integer, String
+from sqlalchemy import Boolean, Column, Enum, Float, Integer, String
 from sqlalchemy.orm import relationship
 
 from app.core.types import IconName
@@ -7,9 +7,10 @@ from app.db.base_class import Base
 
 class OutOfStockItem(Base):
     id = Column(Integer, primary_key=True, nullable=False)
-    buy_or_sell = Column(Boolean, nullable=False) # True means buy, False means sell
+    # True means buy, False means sell
+    buy_or_sell = Column(Boolean, nullable=False)
     name = Column(String, nullable=False)
     icon = Column(Enum(IconName), nullable=False)
-    sell_price = Column(Integer, nullable=True)
+    sell_price = Column(Float, nullable=True)
 
     outofstocks = relationship("OutOfStock", back_populates="item")
